+++ src/main/kotlin/dev/aaronhowser/mods/excessive_utilities/packet/ModPacketHandler.kt
@@ -3,6 +3,7 @@ package dev.aaronhowser.mods.excessive_utilities.packet
 import dev.aaronhowser.mods.aaron.packet.AaronPacketRegistrar
 import dev.aaronhowser.mods.excessive_utilities.packet.client_to_server.ConfigureEnderFrequencyPacket
 import dev.aaronhowser.mods.excessive_utilities.packet.client_to_server.UpdateControlsPacket
+import dev.aaronhowser.mods.excessive_utilities.packet.server_to_client.TransferNodeParticlesPacket
 import dev.aaronhowser.mods.excessive_utilities.packet.server_to_client.UpdateGridPowerPacket
 import net.neoforged.neoforge.network.event.RegisterPayloadHandlersEvent
 
@@ -11,6 +12,11 @@ object ModPacketHandler : AaronPacketRegistrar {
 	fun registerPayloads(event: RegisterPayloadHandlersEvent) {
 		val registrar = event.registrar("1")
 
+		toClient(
+			registrar,
+			TransferNodeParticlesPacket.TYPE,
+			TransferNodeParticlesPacket.STREAM_CODEC
+		)
 		toClient(
 			registrar,
 			UpdateGridPowerPacket.TYPE,